+++ apps/dashboard/src/app/(main)/(protected)/projects/[projectId]/layout.tsx
@@ -1,13 +1,24 @@
+import { SiteLoadingIndicator } from "@/components/site-loading-indicator";
+import { UrlPrefetcher } from "@/lib/prefetch/url-prefetcher";
+import { Suspense } from "react";
 import SidebarLayout from "./sidebar-layout";
 import { AdminAppProvider } from "./use-admin-app";
+export { generateStaticParams } from "@/lib/generate-empty-static-params";
 
-export default async function Layout(
-  props: { children: React.ReactNode, params: Promise<{ projectId: string }> }
+export default function Layout(
+  props: { children: React.ReactNode, modal?: React.ReactNode }
 ) {
   return (
-    <AdminAppProvider projectId={(await props.params).projectId}>
-      <SidebarLayout projectId={(await props.params).projectId}>
-        {props.children}
+    <AdminAppProvider>
+
+      {/* Pre-fetch the current URL to prevent request waterfalls */}
+      <UrlPrefetcher href="" />
+
+      <SidebarLayout>
+        <Suspense fallback={<SiteLoadingIndicator />}>
+          {props.children}
+          {props.modal}
+        </Suspense>
       </SidebarLayout>
     </AdminAppProvider>
   );