+++ apps/dashboard/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @stackframe/stack-dashboard
 
+## 2.8.44
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.44
+  - @stackframe/stack-ui@2.8.44
+  - @stackframe/stack@2.8.44
+
+## 2.8.43
+
+### Patch Changes
+
+- @stackframe/stack@2.8.43
+- @stackframe/stack-shared@2.8.43
+- @stackframe/stack-ui@2.8.43
+
 ## 2.8.42
 
 ### Patch Changes