+++ apps/mock-oauth-server/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @stackframe/mock-oauth-server
 
+## 2.8.44
+
+## 2.8.43
+
 ## 2.8.42
 
 ## 2.8.41