+++ examples/e-commerce/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @stackframe/e-commerce-demo
 
+## 2.8.44
+
+### Patch Changes
+
+- @stackframe/stack@2.8.44
+
+## 2.8.43
+
+### Patch Changes
+
+- @stackframe/stack@2.8.43
+
 ## 2.8.42
 
 ### Patch Changes