+++ examples/react-example/CHANGELOG.md
@@ -1,5 +1,17 @@
 # react-example
 
+## 2.8.44
+
+### Patch Changes
+
+- @stackframe/react@2.8.44
+
+## 2.8.43
+
+### Patch Changes
+
+- @stackframe/react@2.8.43
+
 ## 2.8.42
 
 ### Patch Changes