+++ packages/stack-sc/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @stackframe/stack-sc
 
+## 2.8.44
+
+## 2.8.43
+
 ## 2.8.42
 
 ### Patch Changes