+++ packages/stack-ui/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @stackframe/stack-ui
 
+## 2.8.44
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.44
+
+## 2.8.43
+
+### Patch Changes
+
+- @stackframe/stack-shared@2.8.43
+
 ## 2.8.42
 
 ### Patch Changes