+++ packages/template/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @stackframe/stack
 
+## 2.8.44
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.44
+  - @stackframe/stack-ui@2.8.44
+  - @stackframe/stack-sc@2.8.44
+
+## 2.8.43
+
+### Patch Changes
+
+- Various changes
+  - @stackframe/stack-sc@2.8.43
+  - @stackframe/stack-shared@2.8.43
+  - @stackframe/stack-ui@2.8.43
+
 ## 2.8.42
 
 ### Patch Changes